requirements.txt
backtrader
pandas
numpy
yfinance
matplotlib
pyyaml
jupyter
pytest
